requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Environment and configuration
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Data validation and serialization
pydantic==2.5.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2